More cleanup. Group probing is now implemented.
For some reason, the user cookie isn't getting set on redirect. This may be solvable,
but it looks like a better approach will be to do this integration as a special
"standard servlet errai security impl" module which can be used instead of
errai-security-picketlink. I will make those changes on Monday.