Changing strategies: now working as a direct errai-security integration (no picketlink)
The previous attempts at integrating standard servlet security with PicketLink (so the servlet
container handles login, logout, and identity management) showed more and more clearly that
the benefits of PicketLink are all but lost when these responsibilities are taken away from it.

This new strategy is to integrate errai-security directly with the security facilities of
a standard servlet container. It's more direct and doesn't appear to eliminate any functionality
compared with the errai+picketlink+standard servlet security approach.